feat: move overflow disclaimer to Listbox
